--- network_analysis/find-connected-features-in-utility-network/settings.gradle
@@ -1,1 +1,0 @@
-rootProject.name = 'find-connected-features-in-utility-network'
