--- network_analysis/find-connected-features-in-utility-network/src/main/java/com/esri/samples/find_connected_features_in_utility_network/FindConnectedFeaturesInUtilityNetworkLauncher.java
@@ -1,11 +1,0 @@
-package com.esri.samples.find_connected_features_in_utility_network;
-
-/**
- * Wrapper required for launching a JavaFX 11 app through Gradle or from a jar.
- */
-public class FindConnectedFeaturesInUtilityNetworkLauncher {
-
-  public static void main(String[] args) {
-    FindConnectedFeaturesInUtilityNetworkSample.main(args);
-  }
-}
