+++ network_analysis/trace-a-utility-network/src/main/java/com/esri/samples/trace_a_utility_network/TraceAUtilityNetworkLauncher.java
@@ -1,0 +1,11 @@
+package com.esri.samples.trace_a_utility_network;
+
+/**
+ * Wrapper required for launching a JavaFX 11 app through Gradle or from a jar.
+ */
+public class TraceAUtilityNetworkLauncher {
+
+  public static void main(String[] args) {
+    TraceAUtilityNetworkSample.main(args);
+  }
+}
