+++ network_analysis/trace-a-utility-network/src/main/java/com/esri/samples/trace_a_utility_network/UtilityTerminalListCell.java
@@ -14,7 +14,7 @@
  *
  */
 
-package com.esri.samples.find_connected_features_in_utility_network;
+package com.esri.samples.trace_a_utility_network;
 
 import javafx.scene.control.ListCell;
